src/modules/product/useCases/listCategories/ListCategoriesUseCase.ts
import { ICategoriesRepository } from "@modules/product/repositories/ICategoriesRepository";
import { Category } from "@prisma/client";

class ListCategoriesUseCase {
  private categoriesRepository: ICategoriesRepository;

  constructor(categoriesRepository: ICategoriesRepository) {
    this.categoriesRepository = categoriesRepository;
  }

  async execute(): Promise<Category[]> {
    const categories = await this.categoriesRepository.list();

    return categories;
  }
}

export { ListCategoriesUseCase };
